+++ tests/filters/test_execution_filter.py
@@ -26,6 +26,19 @@ class TestExecutionFilter(unittest.TestCase):
             self.assertEqual(alternative_name, "low_on_steam_wasting_my_time__Importing importer@2023-09-06T01:23:45")
             alternatives = {item["name"] for item in db_map.mapped_table("alternative").valid_values()}
             self.assertIn(alternative_name, alternatives)
+            scenarios = {item["name"] for item in db_map.mapped_table("scenario").valid_values()}
+            self.assertEqual(scenarios, {"low_on_steam", "wasting_my_time"})
+            scenario_alternatives = {
+                (item["scenario_name"], item["alternative_name"], item["rank"])
+                for item in db_map.mapped_table("scenario_alternative").valid_values()
+            }
+            self.assertEqual(
+                scenario_alternatives,
+                {
+                    ("low_on_steam", "low_on_steam_wasting_my_time__Importing importer@2023-09-06T01:23:45", 1),
+                    ("wasting_my_time", "low_on_steam_wasting_my_time__Importing importer@2023-09-06T01:23:45", 1),
+                },
+            )
 
 
 if __name__ == "__main__":